+++ lib/src/bd_level.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/foundation.dart';
+import 'package:meta/meta.dart';
 
 /// An Logging Level of importance of a log message.
 @immutable