+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Provide the functionality to log events.
+* Provide two out-of-the-box log handler (console, file).
+* Provide the feature of formatting logs that allow to format a log record to a pleasant format.
+* Provide a default log formatter.